add relation product and stock

+++ models/stock.go
@@ -4,5 +4,5 @@ type Stock struct {
 	Id 			uint		`gorm:"primaryKey" form:"id"`
 	Amount 		int64		`form:"amount" validate:"required"`
 	UpdatedAt 	int			`gorm:"autoUpdateTime"` 
-	ProductId 	int64		`form:"product_id" validate:"required"`
+	ProductId 	uint		`form:"product_id" validate:"required"`
 }
